pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-gtm-server"
version = "1.0.0"
description = "MCP server for Google Tag Manager integration with Claude"
authors = [
    {name = "MCP GTM Server", email = "noreply@example.com"},
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.5.0",
    "google-auth>=2.0.0",
]

[project.scripts]
mcp-gtm-server = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "test_*.py",
    "__pycache__",
    "*.pyc",
    ".git",
    ".venv",
    ".pytest_cache"
]

[project.urls]
Homepage = "https://github.com/example/mcp-gtm-server"
Repository = "https://github.com/example/mcp-gtm-server"
Issues = "https://github.com/example/mcp-gtm-server/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]